requirements.txt
Flask==2.3.3
Pillow==10.0.1
opencv-python==4.8.1.78
scikit-image==0.21.0
imutils==0.5.4
numpy==1.24.3

# WSGI server for production
gunicorn==21.2.0





